feat: PR 코드리뷰 반영
- 컴포넌트 별 파일 구조 분리
 - 예로 Button.tsx 에 대해
 - Button.css.ts
 - Button.util.ts 혹은 Button.hooks.ts
 - index.ts

 - PostList.tsx 로직 분리
  - 데이터 패칭 로직을 훅으로 분리했습니다.

- 필요없는 scss 파일 삭제
 - global css를 위한 common.css 를 제외하고 삭제했습니다.
 - 종속성으로 인한 merge conflict가 발생하지 않도록 일단 sass 종속은 지우지 않았습니다.

 - Styled Component Naming Convention 은 다음 PR로 반영하겠습니다.
