+++ src/components/Post/PostListNav/PostListNav.hooks.ts
@@ -1,0 +1,1 @@
+export const usePostListNav = () => {};
