--- src/components/Post/index.ts
@@ -1,3 +1,0 @@
-export * from './PostListNav';
-export * from './PostList';
-export * from './PostDetail';
