+++ src/components/commons/Icon/index.ts
@@ -1,3 +1,4 @@
-export * from './icon.type';
-export * from './icon.util';
+export * from './Icon.type';
+export * from './Icon.util';
 export * from './ICON_SET';
+export * from './Icon';
