+++ src/components/layout/FlexBox/FlexListItem.tsx
@@ -1,13 +1,9 @@
 import React, { LiHTMLAttributes } from 'react';
-import FlexBoxCreator from './flex.util';
-import styled from '@emotion/styled';
-import { log } from 'console';
+import FlexBoxCreator from './Flex.util';
 
 export interface FlexListItemProps extends LiHTMLAttributes<HTMLLIElement> {}
 
 export const FlexListItem = ({ children, ...args }: FlexListItemProps) => {
-  // console.log(args);
-
   return <FlexLi {...args}>{children}</FlexLi>;
 };
