+++ src/pages/board/post/[id]/index.tsx
@@ -1,20 +1,11 @@
-import { PostList, PostListNav } from '@/components/Post';
-import { GetServerSideProps } from 'next';
-import { useRouter } from 'next/router';
 import React from 'react';
 
-const PostPage = () => {
-  const router = useRouter();
-  const id = router.query.id;
-
+const Post = () => {
   return (
-    <div>
-      <PostList></PostList>
-      <PostListNav></PostListNav>
-    </div>
+    <main>
+      <section></section>
+    </main>
   );
 };
 
-export default PostPage;
-
-export const getServerSideProps: GetServerSideProps = () => {};
+export default Post;
