--- src/pages/board/post/detail/index.tsx
@@ -1,12 +1,0 @@
-import { PostList, PostListNav } from '@/components/Post';
-import React from 'react';
-
-const Post = () => {
-  return (
-    <main>
-      <section></section>
-    </main>
-  );
-};
-
-export default Post;
