+++ src/styles/config.css
@@ -1,0 +1,14 @@
+* {
+  margin: 0;
+  padding:0;
+  outline:0;
+  appearance:none;
+  border:0;
+  box-sizing: border-box;
+}
+
+body {
+  width: 100vw;
+  height: 100vh;
+  padding-top: 64px;
+}
