--- src/styles/config.scss
@@ -1,17 +1,0 @@
-// 프로젝트 임시 컬러 프로필
-  $color-primary: #6c63ff;
-  $color-success: #00bf8e;
-  $color-warning: #f7c94b;
-  $color-danger: #f75842;
-  $color-danger-variant: rgba(247, 88, 66, 0.4);
-  $color-white: #fff;
-  $color-light: rgba(255,255, 255, 0.7);
-  $color-black: #000;
-  $color-bg: #242e54;
-  $color-bg1: #2e3267;
-  $color-bg2: #424890;
-  
-  // 트랜지션 설정
-  $transition: all 400ms ease;
-  
-
